pyproject.toml
[tool.poetry]
name = "PersonaRAG"
version = "0.1.0"
packages = [{include = "persona_rag"}]
description = ""
authors = ["Saber Zerhoudi <saber.zerhoudi@uni-passau.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.28.0"
aiohttp = "^3.10.5"
numpy = "^1.21.2"
pytest = "^8.2.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.8.4"
nltk = "^3.9.1"
textstat = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
